pyproject.toml
[project]
name = "sltasks"
version = "0.1.0"
description = "Terminal-based Kanban TUI for markdown task management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "textual>=0.89.0",
    "pydantic-settings>=2.6.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
sltasks = "sltasks.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sltasks"]

[tool.hatch.build.targets.wheel.force-include]
"src/sltasks/data/templates" = "sltasks/data/templates"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "textual-dev>=1.8.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "pymdown-extensions>=10.15",
]
